test: move spark test back to dsl_pipeline class
